+++ crates/papyrus_load_test/src/lib.rs
@@ -1,1 +1,34 @@
 pub mod gateway_functions;
+pub mod transaction_functions;
+
+use goose::goose::{GooseUser, TransactionError};
+use serde::de::DeserializeOwned;
+use serde::Deserialize;
+use serde_json::json;
+type MethodResult<T> = Result<T, Box<TransactionError>>;
+
+async fn post_jsonrpc_request<T: DeserializeOwned>(
+    user: &mut GooseUser,
+    method: &str,
+    params: serde_json::Value,
+) -> MethodResult<T> {
+    let request = jsonrpc_request(method, params);
+    let response = user.post_json("", &request).await?.response.map_err(|e| Box::new(e.into()))?;
+    #[derive(Deserialize)]
+    struct TransactionReceiptResponse<T> {
+        result: T,
+    }
+    let response: TransactionReceiptResponse<T> =
+        response.json().await.map_err(|e| Box::new(e.into()))?;
+
+    Ok(response.result)
+}
+
+fn jsonrpc_request(method: &str, params: serde_json::Value) -> serde_json::Value {
+    json!({
+        "jsonrpc": "2.0",
+        "id": "0",
+        "method": method,
+        "params": params,
+    })
+}
